Создаёт Smart-компонент
